06_token/backend/requirements.txt
django
djangorestframework
markdown
django-filter
django-cors-headers
dj-rest-auth
